feat(observability): per-service p95 latency alarms
The only latency alarm is the aggregate ALB p95, which averages out a single
slow journey (one service slow while the ALB-wide p95 looks fine). Add
per-service p95 latency alarms for the six request-serving services from a log
metric filter over the shared middleware's latency_ms (excluding /health).

Also fix shared::middleware to log latency_ms as a u64 number instead of u128
(tracing renders u128 via Debug as a quoted string), so the numeric metric
filter can extract it. Requires redeploying all services for the filter to
start receiving numeric values.

Verified: clippy -D warnings + cargo test -p shared; aws logs
test-metric-filter confirms the pattern matches responses but not /health or
non-response logs; terraform fmt + validate.

+++ services/shared/src/middleware.rs
@@ -31,7 +31,7 @@ pub fn with_logging<S: Clone + Send + Sync + 'static>(app: Router<S>) -> Router<
                     |res: &axum::http::Response<_>, latency: std::time::Duration, _span: &Span| {
                         tracing::info!(
                             status = res.status().as_u16(),
-                            latency_ms = latency.as_millis(),
+                            latency_ms = latency.as_millis() as u64,
                             "response"
                         );
                     },
@@ -40,7 +40,7 @@ pub fn with_logging<S: Clone + Send + Sync + 'static>(app: Router<S>) -> Router<
                     |err: ServerErrorsFailureClass, latency: std::time::Duration, _span: &Span| {
                         tracing::error!(
                             error = %err,
-                            latency_ms = latency.as_millis(),
+                            latency_ms = latency.as_millis() as u64,
                             "request failed"
                         );
                     },
